backend/requirements.txt
djangorestframework
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
gunicorn>=19.7.1
pylint==2.4.4
pylint-django
callee
coverage
drf-yasg
reportlab
bcrypt
pyjwt
python-dateutil
